rag_llms/requirements.txt
pypdf==5.6.1
python-dotenv==1.1.1

langchain==0.3.26
langchain-community==0.3.26
langchain-openai==0.3.25
langchain-chroma==0.2.4
langchain-ollama==0.3.3

huggingface-hub==0.33.1
sentence-transformers==4.1.0